New functionality: dot-product
Tests in terms of performances have been made to compare NumPy dot and MathModule dot.

Results show that the kernel is more efficient for large input-vectors.
Unfortunately, the time consumed by allocating and copying memory to GPU is too high and total performance is worse using GPU than CPU alone - on the tests.